RAG_system/flask_backend/requirements.txt
docling==2.23.1
docling_core==2.29.0
Flask==3.1.1
flask_cors==5.0.1
gunicorn==21.2.0
huggingface_hub==0.28.1
jieba==0.42.1
ollama==0.5.1
openai==1.84.0
pandas==2.2.3
psutil==5.9.0
PyStemmer==2.2.0.3
qdrant_client==1.13.3
sentence_transformers==3.4.1
transformers==4.42.4
typing_extensions==4.14.0
Werkzeug==3.1.3
python-dotenv==1.0.0
